Create folders and organize files
