+++ src/App.scss
@@ -1,3 +1,5 @@
+@use './Styles/reset';
+
 main {
   background-color: black;
   color: white;
