+++ src/Components/AllMovies/AllMovies.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import MovieCard from './MovieCard';
+import MovieCard from '../MovieCard/MovieCard';
 import './AllMovies.scss';
 
 const AllMovies = ({movies}) => {
