+++ src/Components/AllMovies/AllMovies.scss
