+++ src/Components/MovieCard/MovieCard.js
@@ -7,6 +7,7 @@ const MovieCard = ({id, title, rating, poster}) => {
             <img className='card-image' src={poster} alt={title} />
             <p>{title}</p>
             <p>{rating.toFixed(1)}</p>
+            <p className="description">This is my movie description.</p>
         </section>
     )
 }
